Battle Structure Setup
Added funtionality to LevelManager and SpawnController, to allow for a string input on buttons that lead to the battle scene to determine what enemies are present. 

battleType can be either "random", which runs the old random spawn functions, or a specific string which is passed through to DetermineBattleType(), where specific enemies are set to spawn.